backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
groq==0.9.0
python-dotenv==1.0.0
httpx==0.25.2